Adds identifier utility methods and tests
Adds `is_zero` and `is_max` methods to the `Identifier` struct for
checking if an identifier is composed of all zeros or all 255s, respectively.

Also adds a utility function `random_identifier_greater_than` to generate
a random identifier that is greater than a given identifier.

Includes corresponding unit tests for the new methods and function.