Simplifies MockNetwork processor registration
Updates the `MockNetwork` to allow safe and concurrent
access to the message processor. This involves changing
the `processor` field to be wrapped in an `Arc>`,
allowing it to be safely shared across threads.
Also modifies the `register_processor` function to take
`&self` instead of `&mut self`, enabling concurrent
registration from different threads. This change avoids
the need to acquire a lock on the entire `MockNetwork`
when registering a processor.